requirements.txt
streamlit
Pillow
rembg
opencv-python-headless
numpy
flask
flask-cors
